LayoutTests/imported/w3c/web-platform-tests/css/css-fonts/parsing/font-shorthand-variant-expected.txt


FAIL font shorthand returns only CSS2 font-variant values assert_equals: font should be empty expected "" but got "titling-caps medium serif"
